refactor: Extract `Modal.Header` into separate component
