+++ src/components/Modal/Header/index.tsx
@@ -1,0 +1,1 @@
+export { Component as Header } from './component';
