+++ src/components/Select/variant/ModalSelect/component.tsx
@@ -4,16 +4,27 @@ import { useBuildTestId } from '../../../../modules/test-ids';
 import { Modal } from '../../../Modal';
 import { SelectContext } from '../../context';
 
-export type Props = React.ComponentPropsWithoutRef<typeof Modal>;
+export type Props = React.ComponentPropsWithoutRef<typeof Modal> & {
+  title?: React.ReactNode;
+  isLoading?: boolean;
+};
 
-export const Component = ({ onClose, 'data-testid': testId, children, ...otherProps }: Props) => {
+export const Component = ({
+  title,
+  children,
+  isLoading,
+  onClose,
+  'data-testid': testId,
+  ...otherProps
+}: Props) => {
   const context = useMemo(() => ({ onClose, testId }), [onClose, testId]);
   const buildTestId = useBuildTestId(testId);
 
   return (
     <SelectContext.Provider value={context}>
       <Modal {...otherProps} onClose={onClose} data-testid={buildTestId()}>
-        <Modal.Content>{children}</Modal.Content>
+        <Modal.Header title={title} isLoading={isLoading} onClose={onClose} />
+        <Modal.Content padding="none">{children}</Modal.Content>
       </Modal>
     </SelectContext.Provider>
   );
